django_test_failing_app/requirements.txt
Django<2.3
psycopg2-binary
factory-boy
